Moved various tests that were in test_dark_monitor to separate test modules based on which modules are being tested.  Also added sphinx api docs for said tests.
