+++ jwql/tests/test_utils.py
@@ -22,7 +22,7 @@ import os
 from pathlib import Path
 import pytest
 
-from jwql.utils.utils import copy_files, get_config, filename_parser
+from jwql.utils.utils import copy_files, get_config, filename_parser, filesystem_path
 
 
 FILENAME_PARSER_TEST_DATA = [
@@ -246,8 +246,8 @@ FILENAME_PARSER_TEST_DATA = [
 
 
 def test_copy_files():
-    """Test that files are copied successfully
-    """
+    """Test that files are copied successfully"""
+
     # Create an example file to be copied
     data_dir = os.path.dirname(__file__)
     file_to_copy = 'file.txt'
@@ -267,7 +267,7 @@ def test_copy_files():
     os.remove(original_file)
     os.remove(copied_file)
 
-    
+
 @pytest.mark.skipif(os.path.expanduser('~') == '/home/jenkins',
                     reason='Requires access to central storage.')
 def test_get_config():
@@ -331,3 +331,15 @@ def test_filename_parser_nonJWST():
     with pytest.raises(ValueError):
         filename = 'not_a_jwst_file.fits'
         filename_parser(filename)
+
+
+@pytest.mark.skipif(os.path.expanduser('~') == '/home/jenkins',
+                    reason='Requires access to central storage.')
+def test_filesystem_path():
+    """Test that a file's location in the filesystem is returned"""
+
+    filename = 'jw96003001001_02201_00001_nrca1_dark.fits'
+    check = filesystem_path(filename)
+    location = os.path.join(get_config()['filesystem'], 'jw96003', filename)
+
+    assert check == location
